# Conflicts:
#	activerecord/lib/active_record/tasks/database_tasks.rb

+++ activerecord/test/cases/scoping/default_scoping_test.rb
@@ -225,6 +225,15 @@ class DefaultScopingTest < ActiveRecord::TestCase
     assert_match(/mentor_id/, reload_sql)
   end
 
+  def test_default_scope_with_all_queries_runs_on_reload_but_default_scope_without_all_queries_does_not
+    Mentor.create!
+    dev = DeveloperWithIncludedMentorDefaultScopeNotAllQueriesAndDefaultScopeFirmWithAllQueries.create!(name: "Eileen")
+    reload_sql = capture_sql { dev.reload }.first
+
+    assert_no_match(/mentor_id/, reload_sql)
+    assert_match(/firm_id/, reload_sql)
+  end
+
   def test_nilable_default_scope_with_all_queries_runs_on_reload
     dev = DeveloperWithDefaultNilableFirmScopeAllQueries.create!(name: "Nikita")
     reload_sql = capture_sql { dev.reload }.first
@@ -400,6 +409,20 @@ class DefaultScopingTest < ActiveRecord::TestCase
     expected = Developer.all.collect(&:name)
     received = Developer.includes(:projects).select(:id).unscope(:includes, :select).collect(&:name)
     assert_equal expected, received
+  end
+
+  def test_unscope_eager_load
+    expected = Developer.all.collect(&:name)
+    received = Developer.eager_load(:projects).select(:id).unscope(:eager_load, :select)
+    assert_equal expected, received.collect(&:name)
+    assert_equal false, received.first.projects.loaded?
+  end
+
+  def test_unscope_preloads
+    expected = Developer.all.collect(&:name)
+    received = Developer.preload(:projects).select(:id).unscope(:preload, :select)
+    assert_equal expected, received.collect(&:name)
+    assert_equal false, received.first.projects.loaded?
   end
 
   def test_unscope_having
@@ -691,39 +714,41 @@ class DefaultScopingTest < ActiveRecord::TestCase
 end
 
 class DefaultScopingWithThreadTest < ActiveRecord::TestCase
-  self.use_transactional_tests = false
+  unless in_memory_db?
+    self.use_transactional_tests = false
 
-  def test_default_scoping_with_threads
-    2.times do
-      Thread.new {
-        assert_includes DeveloperOrderedBySalary.all.to_sql, "salary DESC"
-        DeveloperOrderedBySalary.connection.close
-      }.join
+    def test_default_scoping_with_threads
+      2.times do
+        Thread.new {
+          assert_includes DeveloperOrderedBySalary.all.to_sql, "salary DESC"
+          DeveloperOrderedBySalary.connection.close
+        }.join
+      end
+    end
+
+    def test_default_scope_is_threadsafe
+      2.times { ThreadsafeDeveloper.unscoped.create! }
+
+      threads = []
+      assert_not_equal 1, ThreadsafeDeveloper.unscoped.count
+
+      barrier_1 = Concurrent::CyclicBarrier.new(2)
+      barrier_2 = Concurrent::CyclicBarrier.new(2)
+
+      threads << Thread.new do
+        Thread.current[:default_scope_delay] = -> { barrier_1.wait; barrier_2.wait }
+        assert_equal 1, ThreadsafeDeveloper.all.to_a.count
+        ThreadsafeDeveloper.connection.close
+      end
+      threads << Thread.new do
+        Thread.current[:default_scope_delay] = -> { barrier_2.wait }
+        barrier_1.wait
+        assert_equal 1, ThreadsafeDeveloper.all.to_a.count
+        ThreadsafeDeveloper.connection.close
+      end
+      threads.each(&:join)
+    ensure
+      ThreadsafeDeveloper.unscoped.destroy_all
     end
   end
-
-  def test_default_scope_is_threadsafe
-    2.times { ThreadsafeDeveloper.unscoped.create! }
-
-    threads = []
-    assert_not_equal 1, ThreadsafeDeveloper.unscoped.count
-
-    barrier_1 = Concurrent::CyclicBarrier.new(2)
-    barrier_2 = Concurrent::CyclicBarrier.new(2)
-
-    threads << Thread.new do
-      Thread.current[:default_scope_delay] = -> { barrier_1.wait; barrier_2.wait }
-      assert_equal 1, ThreadsafeDeveloper.all.to_a.count
-      ThreadsafeDeveloper.connection.close
-    end
-    threads << Thread.new do
-      Thread.current[:default_scope_delay] = -> { barrier_2.wait }
-      barrier_1.wait
-      assert_equal 1, ThreadsafeDeveloper.all.to_a.count
-      ThreadsafeDeveloper.connection.close
-    end
-    threads.each(&:join)
-  ensure
-    ThreadsafeDeveloper.unscoped.destroy_all
-  end
-end unless in_memory_db?
+end
